Moved error handlers to separate file
